+++ src/test/resources/sample/Sample2.java
@@ -1,0 +1,8 @@
+package sample;
+
+public class Sample2 {
+
+    public String foo() {
+        return "test";
+    }
+}