+++ src/test/resources/sample/SampleTest.groovy
@@ -1,0 +1,12 @@
+package sample
+
+import spock.lang.Specification
+
+class SampleTest extends Specification{
+
+    def "Check sample"() {
+
+        expect:
+        new Sample().foo() == 'test'
+    }
+}