+++ src/test/resources/sample/SampleTest2.groovy
@@ -1,0 +1,12 @@
+package sample
+
+import spock.lang.Specification
+
+class SampleTest2 extends Specification{
+
+    def "Check sample"() {
+
+        expect:
+        new Sample2().foo() == 'test'
+    }
+}